refactor(socket): consolidate maxAttachments override into a single cypressParser
Per Bugbot review, the UnlimitedAttachmentsDecoder + cypressParser
pair was duplicated across browser.ts and node/socket.ts, and a third
inline maxAttachments override sat in utils.ts. Move the canonical
definition into utils.ts (which is already shared between the browser
and node tsconfigs) and import `cypressParser` from both browser.ts
and socket.ts. utils.ts itself uses `new cypressParser.Decoder()`
instead of inlining the option.

Single source of truth: if the policy ever changes (different cap,
new option), there's now one place to update.
